fix array and resolve merge conflicts
